requirements.txt
psycopg2-binary==2.9.9
gspread==6.1.4
google-auth==2.29.0
